Push scheduled command test env. to the sub-package level
